rename run -> parameter; add docs
